requirements.txt
gradio
python-dotenv
langchain
langchain-groq